requirements.txt
langchain~=0.3.27
langchain-core
langchain-community
langchain-text-splitters
sentence-transformers
pypdf
docx2txt
unstructured
fastapi~=0.116.1
uvicorn
streamlit
python-dotenv~=1.1.1
pinecone~=7.3.0
langchain-pinecone
langchain-groq
langchain-tavily
requests
uuid
langchain-huggingface
pydantic~=2.11.7

torch~=2.8.0
dnspython~=2.7.0
email-validator~=2.3.0


